pyproject.toml
[project]
name = "bbc-audio-scraper"
version = "0.1.0"
description = "Download BBC audio programmes, transcribe with Whisper (free), and chat with transcripts using Google AI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gradio>=4.0.0",
    "openai-whisper",
    "google-generativeai",
    "beautifulsoup4",
    "requests",
    "feedparser",
    "chromadb",
    "langchain",
    "langchain-google-genai",
    "langchain-community",
    "pydub",
    "numpy",
    "pandas",
    "python-dotenv",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]
